\name{RecAssoRules-package}
\alias{RecAssoRules}
\docType{package}
\title{Recursive mining for frequent itemSet and confident association rules.}
\description{This first version makes it possible to carry out the usual applications of association rule Mining.This by using recursive algorithms. All of the functions are using the C++ language. For more usage security, it mainly uses the std and stl libraries, reducing only very slightly the efficiency of the algorithms. It will be updated soon with functions allowing it to be applied on non-fixed databases, as well as a more practical interface.}


