\name{Transacfruits}
\alias{Transacfruits}
\docType{data}
\title{Transacfruits Data Set}
\description{
  The Transacfruits dataset is a synthetic database. It thus represents fictitious transactions, simulating fruit purchases, among a random list.
}
\usage{
data("Transacfruits")
}

\details{
  Transacfruits is made up of 1 columns and 100 rows. There are therefore 100 transactions. For each of them one or more fruits were purchased from a list of 15 fruits. It was generated by binomial simulation, for all transactions, there was for each fruit a fixed probability p of being bought.
    }
